% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sas_to_xpt.R
\name{convert_sas_to_xpt}
\alias{convert_sas_to_xpt}
\title{convert sas to xpt function}
\usage{
convert_sas_to_xpt(sas_folder)
}
\arguments{
\item{sas_folder}{input a file directory where your sas files are}
}
\value{
xpt files
}
\description{
convert sas to xpt function
}
\examples{
xpt_folder <- 'XX:/convert/xpt_test'
xpt_files <- list.files(path = xpt_folder, pattern = "\\\\.xpt$", full.names = TRUE)
for (file in xpt_files) {
sas_file <- gsub("\\\\.xpt$", ".sas7bdat", file)
if (!file.exists(sas_file)) {
  df <- haven::read_xpt(file)
  haven::write_sas(df, sas_file)
  cat("Converted", file, "to SAS7BDAT format.\n")
} else {
  cat(sas_file, "already exists. Skipping...\n")
  }
  }
}
